+++ apps/forms/layouts/Main/Navigation/styles.ts
@@ -1,6 +1,56 @@
 import styled, { createGlobalStyle, css } from "styled-components";
-import { scrollbar } from "@basestack/design-system/styles";
+import { space, SpaceProps } from "styled-system";
 import { rem, transparentize } from "polished";
+import { scrollbar } from "@basestack/design-system/styles";
+
+// Desktop Navigation
+export const DesktopContainer = styled.nav`
+  position: sticky;
+  top: 0;
+  z-index: ${({ theme }) => theme.zIndex.navigation};
+  display: flex;
+  align-items: center;
+  flex-shrink: 0;
+  background-color: ${({ theme }) =>
+    theme.colors[theme.isDarkMode ? "gray800" : "white"]};
+  box-shadow: ${({ theme }) => theme.shadow.elevation2};
+  height: ${rem("64px")};
+  padding: 0 ${({ theme }) => theme.spacing.s5};
+`;
+
+const desktopFlexCenter = css`
+  display: flex;
+  align-items: center;
+`;
+
+export const DesktopList = styled.ul<SpaceProps>`
+  ${space};
+  ${desktopFlexCenter};
+`;
+
+export const DesktopListItem = styled.li<SpaceProps>`
+  ${space};
+`;
+
+export const LogoContainer = styled.div`
+  ${desktopFlexCenter};
+  position: relative;
+  padding-right: ${({ theme }) => theme.spacing.s4};
+  margin-right: ${({ theme }) => theme.spacing.s1};
+  cursor: pointer;
+
+  &::before {
+    content: "";
+    position: absolute;
+    height: 20px;
+    right: -1px;
+    width: 2px;
+    background-color: ${({ theme }) =>
+      theme.colors[theme.isDarkMode ? "gray600" : "gray100"]};
+  }
+`;
+
+// Mobile Navigation
 
 export const GlobalStyle = createGlobalStyle`
   body {
