+++ apps/forms/layouts/Main/index.tsx
@@ -1,21 +1,15 @@
-import React, { Fragment, useEffect, useState } from "react";
-import { useMedia } from "react-use";
-import { useTheme } from "styled-components";
+import React, { Fragment } from "react";
 // Router
 import { useRouter } from "next/router";
 // Auth
 import { useSession } from "next-auth/react";
 // Components
 import { Splash, Loader } from "@basestack/design-system";
-import NavigationDrawer from "components/Navigation/Mobile";
-import Navigation from "components/Navigation/Desktop";
+import Navigation from "./Navigation";
 // Server
 import { trpc } from "libs/trpc";
 
 const MainLayout = ({ children }: { children: React.ReactNode }) => {
-  const theme = useTheme();
-  const [isDrawerOpen, setIsDrawerOpen] = useState(false);
-  const isMobile = useMedia(theme.device.max.lg, false);
   const router = useRouter();
   const { status } = useSession({
     required: true,
@@ -23,12 +17,6 @@ const MainLayout = ({ children }: { children: React.ReactNode }) => {
       router.push("/auth/sign-in");
     },
   });
-
-  useEffect(() => {
-    if (!isMobile) {
-      setIsDrawerOpen(false);
-    }
-  }, [isMobile]);
 
   const { data, isLoading: isLoadingForms } = trpc.form.all.useQuery(
     undefined,
@@ -58,16 +46,7 @@ const MainLayout = ({ children }: { children: React.ReactNode }) => {
 
   return (
     <Fragment>
-      <Navigation
-        onClickMenuButton={() => setIsDrawerOpen(true)}
-        isDesktop={!isMobile}
-        data={data}
-      />
-      <NavigationDrawer
-        data={data}
-        onClose={() => setIsDrawerOpen(false)}
-        isDrawerOpen={isDrawerOpen}
-      />
+      <Navigation data={data} />
       {children}
     </Fragment>
   );
